Move mocked userset to vmess/protocol/user
