+++ proxy/vmess/protocol/vmess_test.go
@@ -9,7 +9,7 @@ import (
 	v2net "github.com/v2ray/v2ray-core/common/net"
 	"github.com/v2ray/v2ray-core/proxy/vmess/config"
 	"github.com/v2ray/v2ray-core/proxy/vmess/protocol/user"
-	"github.com/v2ray/v2ray-core/testing/mocks"
+	"github.com/v2ray/v2ray-core/proxy/vmess/protocol/user/testing/mocks"
 	"github.com/v2ray/v2ray-core/testing/unit"
 )
